backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# OpenAI
openai==1.10.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utils
python-dotenv==1.0.0
email-validator==2.1.0
python-dateutil==2.8.2

# Task Queue (optional for async processing)
celery==5.3.4
redis==5.0.1

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
isort==5.13.2
